+++ src/views/seq/series.vue
@@ -39,8 +39,10 @@
         </div>
 
         <vxe-grid
+          highlight-current-row
           highlight-hover-row
           auto-resize
+          height="570"
           :ref="seriesTable.ref"
           :loading="seriesTable.loading"
           :columns="seriesTable.columns"
@@ -53,7 +55,7 @@
         </vxe-grid>
       </a-layout-content>
 
-      <a-layout-sider width="300">
+      <a-layout-sider width="350">
         <span style="line-height:32px;">引物</span>
         <div class="table-operator">
           <a-button-group>
@@ -64,6 +66,7 @@
         <vxe-grid
           highlight-hover-row
           auto-resize
+          height="570"
           :ref="seriesPrimersTable.ref"
           :loading="seriesPrimersTable.loading"
           :columns="seriesPrimersTable.columns"
@@ -74,19 +77,31 @@
       </a-layout-sider>
     </a-layout>
 
+    <a-modal
+      title="选择引物"
+      width="1000px"
+      :visible="choosePrimer.visible"
+      :footer="null"
+      @cancel="choosePrimer.visible = false">
+      <choose-primer @callback="setPrimer"></choose-primer>
+    </a-modal>
   </div>
 </template>
 
 <script>
+import ChoosePrimer from '@/components/seq/choosePrimer';
+
 export default {
   name: 'Series',
   components: {
+    ChoosePrimer
   },
   data () {
     return {
       form: this.$form.createForm(this),
       queryParam: {},
       seriesTable: {
+        child: 'seriesPrimersTable',
         id: 0,
         ref: 'seriesTable',
         xTable: null,
@@ -107,13 +122,20 @@ export default {
         }
       },
       seriesPrimersTable: {
+        parent: 'seriesTable',
         id: 0,
         ref: 'seriesPrimersTable',
         xTable: null,
+        editIndex: -1,
+        editData: null,
         loading: false,
         tableData: [],
         columns: [],
         editRules: {}
+      },
+      choosePrimer: {
+        visible: false,
+        formData: {}
       }
     };
   },
@@ -126,13 +148,15 @@ export default {
     // 设置表格列属性
     setColumn () {
       const tableName = 'seriesTable';
-      const { formatter } = this.$units;
+      const { formatter } = this.$utils;
       const { basic } = this.$store.state;
 
       const columns = [
+        // { type: 'radio', width: 40 },
+        { type: 'selection', width: 40 },
         { type: 'index', width: 40 },
         { title: '编号', field: 'code' },
-        { title: '名称', field: 'name', editRender: { name: 'input' } },
+        { title: '名称', field: 'name', editRender: { name: 'AInput' } },
         { title: '状态', field: 'status', formatter: function ({ cellValue }) { return formatter(basic.status, cellValue); } },
         { title: '创建人', field: 'creatorName' },
         { title: '创建时间', field: 'createDate' },
@@ -197,7 +221,7 @@ export default {
         this[tableName].pagerConfig.currentPage = params.page;
         this[tableName].pagerConfig.pageSize = params.rows;
 
-        this[tableName].editIndex = -1;
+        // this[tableName].editIndex = -1;
       }).finally(() => {
         this[tableName].loading = false;
       });
@@ -212,8 +236,11 @@ export default {
         id: --this[tableName].id
       };
 
-      this[tableName].tableData = [newData, ...this[tableName].tableData];
-      table.setActiveRow(newData);
+      // this[tableName].tableData = [newData, ...this[tableName].tableData];
+      // table.setActiveRow(newData);
+      table.insert(newData).then(({ row }) => {
+        table.setActiveRow(row);
+      });
       this[tableName].editIndex = 0;
     },
     // 修改
@@ -242,19 +269,21 @@ export default {
         });
       }
     },
-    // 退出编辑
-    handleQuitEdit ({ row, rowIndex, tableName, xTable }) {
+    /**
+     * 退出编辑
+     * status 有值，退出修改操作，还原数据
+     *        无值，退出新增操作，删除新增的行
+     */
+    handleQuitEdit ({ row, xTable }) {
       xTable.clearActived().then(() => {
-        this[tableName].editIndex = -1;
-        if (!row.status) {
-          this[tableName].tableData.splice(rowIndex, 1);
+        if (typeof row.status === 'number') {
+          xTable.revert(row);
         } else {
-          this.$set(this[tableName].tableData, rowIndex, this[tableName].editData);
-          this[tableName].editData = null;
+          xTable.remove(row);
         }
       });
     },
-    // 点击载体表格时
+    // 点击表格行时
     handleCellClick ({ row }) {
       const tableName = 'seriesPrimersTable';
       if (!row.id || row.id < 0) {
@@ -282,11 +311,11 @@ export default {
     // 为系列之引物设置列
     setColumnToPrimer () {
       const tableName = 'seriesPrimersTable';
-      const { formatter } = this.$units;
+      const { formatter } = this.$utils;
       const { seq } = this.$store.state;
       const columns = [
         { title: '引物编号', field: 'code' },
-        { title: '引物名称', field: 'name', editRender: { name: 'AInput' } },
+        { title: '引物名称', field: 'name', editRender: { name: 'SInputSearch', events: { search: this.selectPrimer } } },
         { title: '引物类型', field: 'type', formatter: function ({ cellValue }) { return formatter(seq.primerType, cellValue); } },
         {
           title: '操作',
@@ -324,13 +353,34 @@ export default {
     // 新增一可编辑行
     handleAddRowToPrimers () {
       const tableName = 'seriesPrimersTable';
-      const table = this.$refs[tableName].$refs.xTable;
+      const table = this[tableName].xTable;
+      const parentTable = this[this[tableName].parent].xTable;
+      console.log(parentTable.getSelectRecords());
       const newData = {
-        id: --this[tableName].id,
-        name: ''
+        id: --this[tableName].id
       };
-      this[tableName].tableData = [newData, ...this[tableName].tableData];
-      table.setActiveRow(newData);
+      table.insert(newData).then(({ row }) => {
+        table.setActiveRow(row);
+      });
+      // this[tableName].editIndex = 0;
+    },
+    // 显示引物选择框
+    selectPrimer () {
+      this.choosePrimer.visible = true;
+    },
+    // 设置引物
+    setPrimer (data) {
+      const tableName = 'seriesPrimersTable';
+      const table = this[tableName].xTable;
+      const primer = {
+        primerId: data.id,
+        code: data.code,
+        name: data.name,
+        type: data.type
+      };
+      Object.assign(table.getInsertRecords()[0], primer);
+
+      this.choosePrimer.visible = false;
     }
   }
 };