+++ vue.config.js
@@ -1,4 +1,5 @@
 const path = require('path');
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
 module.exports = {
   // 部署应用的基本URL
@@ -9,6 +10,11 @@ module.exports = {
       alias: {
         'vue$': path.resolve(__dirname, 'node_modules/vue/dist/vue.js')
       }
+    }
+  },
+  chainWebpack: config => {
+    if (process.env.use_analyzer) {
+      config.plugin('webpack-bundle-analyzer').use(BundleAnalyzerPlugin);
     }
   },
   // webpack devServer